Add a WithExisting<Rel> function to one to many relations
Currently the factory code forcefully sets a template for required
relations if one isn't specified manually, even if the using code
manually specifies IDs for foreign keys.

With optional relations, it is possible to specify existing already
persisted records by setting the ID instead of the relation in a
template. With required relations, this is impossible because of the
above.

This commit lets users to select existing entries to one to many relations
with specifying an existing record.

The commit also introduces related generic functions, for example now it
is possible to use the FromExisting<Relation> functions in the main
Factory class, and return templates that are based on existing records,
and then can be customized further.